requirements.txt
fastapi==0.116.1
uvicorn==0.35.0

# Beanie ODM (pydantic v1 compatible)
beanie==1.18.0

# MongoDB drivers (compatible with Beanie 1.18.0)
pymongo==4.7.2
motor==3.3.2

# Validation and env management
pydantic==1.10.14
python-dotenv==1.0.1
python-decouple==3.8

# Existing project deps
boto3==1.35.43
sendgrid==6.12.4
PyJWT==2.8.0
requests==2.31.0
